requirement.txt
numpy==1.18.5
opencv-python==4.2.0.34
pandas==1.0.4
paramiko==2.7.1
psutil==5.6.2
Pillow==7.1.2
PyInstaller==3.6
pytest==5.4.3
python-dateutil==2.8.0
PyYAML==3.12
scapy==2.4.3
scapy-python3==0.26
scipy==1.4.1
uiautomation==2.0.5
urllib3==1.22
xlrd==1.2.0
numba==0.49.1
openpyxl==3.0.3
pyuserinput==0.1.11
mss==5.1.0
python-xlib==0.27
pyautogui==0.9.50
requests==2.9.1
jinja2==2.10.1
markupsafe==1.1.1